some tests to SchedulingRestTest class added

+++ timesheet/src/main/java/org/catrobat/jira/timesheet/services/TimesheetService.java
@@ -51,12 +51,13 @@ public interface TimesheetService {
      * @param targetHoursTheory   specifies the amount of hours the user has to
      *                            invest in theoretical work
      * @param lectures            describes the lecture in which the user is enrolled
+     * @param ects
      * @return the new Timesheet
      */
     @Nonnull
     Timesheet add(String userKey, int targetHoursPractice, int targetHoursTheory,
                   int targetHours, int targetHoursCompleted, int targetHoursRemoved, String lectures,
-                  String reason, int ects, String latestEntryDate, Boolean isActive, Boolean isEnabled,
+                  String reason, double ects, String latestEntryDate, Boolean isActive, Boolean isEnabled,
                   Boolean isMasterThesisTimesheet);
 
     /**
